+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ensembles - zoom on all models
 - Fixed colors so that color of mono-element entity (e.g. Na, Mg) equals color of the element
 - Added ball-and-stick visual for non-standard residues (so far without element colors)
+- Opacity in entity images is dependent on structure size
 
 ## [1.0.0] - 2023-06-20